+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRZ9VK21SQ4FRCYWRR3VW0NW"
+        "01KRZKVD44PV9HKRF44AS9P47C"
     )
     r"""The unique identifier of the data source"""