+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRZ9VK1ZGGV9J8WVVVNT7KEF"
+        "01KRZKVD40169ADRAGD3ZNAZE0"
     )
     r"""The unique identifier of the data source"""