+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRZ9VK215BENCXR2YFYFC5BT"
+        "01KRZKVD42ZRKH7921W2ZB3G2P"
     )
     r"""The unique identifier of the data source"""