+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRZ9VK23FZQQPZEFDKCHZZYS"
+        "01KRZKVD48E9PA83WDW5YMWK6G"
     )
     r"""The unique identifier of the data source"""