+++ packages/orq-rc/RELEASES.md
@@ -7999,3 +7999,13 @@ Based on:
 - [python v4.10.0-rc.13] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.13] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.13 - packages/orq-rc
+
+## 2026-05-19 08:00:50
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.1 (2.884.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.14] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.14] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.14 - packages/orq-rc